Scoring added [edit]
Scoring was added in the following way. First, we detect which players are firing. Then we detect the changing of the health status of a vehicle, if it's decreasing, we check which players are firing at the moment. Then we check which of those players are in the vicinity of the player taking damage. Then we further filter out players by checking if the players that are firing and are in the vicinity of the victim are also headed towards the victim. The latter is done so that you don't just fly close to a player that is taking damage, hold your fire button and getting points without actually hitting the victim.

[EDIT]
In the previous commit I forgot to filter out positive health changes. This is done now. The reason is that if you would be flying close to a player that picks up a health pickup while firing you would get a negative score from that, since the added score is equal to the negative of the changed amount of health.